\name{abind3DAlong2}
\alias{abind3DAlong2}
\title{abind3DAlong2}
\description{bind a list of 3D arrays along the second dimension}
\usage{abind3DAlong2(Xs)}
\arguments{
  \item{Xs}{list with matrices}
}
\details{The dimensionnames correspond to that of Xs[[1]], except that the columnnames are NULL}


\author{Thomas Wutzler}





