\name{abind4DAlong3}
\alias{abind4DAlong3}
\title{abind4DAlong3}
\description{bind a list of 4D arrays along the third dimension}
\usage{abind4DAlong3(Xs)}
\arguments{
  \item{Xs}{list with matrices}
}
\details{The dimensionnames correspond to that of Xs[[1]], except the names of the third dimension are NULL}


\author{Thomas Wutzler}





