\name{as.POSIXctz}
\alias{as.POSIXctz}
\title{as POSIXctz}
\description{convert to POSIXctz}
\usage{as.POSIXctz(x, ...)}
\arguments{
  \item{x}{object that can be coerced by \code{\link{as.POSIXct}}}
  \item{\dots}{further arguments to \code{\link{as.POSIXct}}}
}
\details{calls \code{\link{as.POSIXct}} and prepends the "POSIXctz" attribute }


\author{Thomas Wutzler}





