\name{catHeadNamedVector}
\alias{catHeadNamedVector}
\title{catHeadNamedVector}
\description{print first numeric vector in form name=value,name2=value2 }
\usage{catHeadNamedVector(x, digits = 3, nmax = 4)}
\arguments{
  \item{x}{
}
  \item{digits}{
}
  \item{nmax}{
}
}



\author{Thomas Wutzler}





