\name{copy2clip}
\alias{copy2clip}
\title{copy2clip}
\description{copies argument to the clipboard}
\usage{
copy2clip(x, col.names = NA, row.names = FALSE, quote = FALSE, ...)
}
\arguments{
  \item{x}{
}
  \item{col.names}{
}
  \item{row.names}{
}
  \item{quote}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





