\name{fileExt}
\alias{fileExt}
\title{fileExt}
\description{Return the file extension}
\usage{
fileExt(filenames)
}
\arguments{
  \item{filenames}{
}
}



\author{Thomas Wutzler}




\examples{
##seealso<< code{\link{fileExt}}, \link{twMisc}
filenames=c("text.txt","dir/some.test.yml","a")
fileExt(filenames)
}
