\name{imageNames}
\alias{imageNames}
\title{imageNames}
\description{imagePlot using row and columnnames as axis.}
\usage{imageNames(x, ...)}
\arguments{
  \item{x}{matrix to plot}
  \item{\dots}{further argumetns to \code{\link{image}}}
}
\details{modified from http://www.phaget4.org/R/image_matrix.html}


\author{Thomas Wutzler}





