\name{isClusterRunning}
\alias{isClusterRunning}
\title{isClusterRunning}
\description{test if given cluster is running}
\usage{isClusterRunning(cl)}
\arguments{
  \item{cl}{cluster to test}
}

\value{TRUE if clusterEval worked, FALSE if not}

\author{Thomas Wutzler}



\seealso{\code{\link{parLapplySave}}, \code{\link{isParallel}}, \code{\link{twMisc}}}

