\name{isParallel}
\alias{isParallel}
\title{isParallel}
\description{isClusterRunning and number of processors != 1 }
\usage{isParallel(cl = NULL)}
\arguments{
  \item{cl}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{parLapplySave}}, \code{\link{isClusterRunning}}, \code{\link{twMisc}}}

