\name{POSIXctz-class}
\Rdversion{1.1}
\docType{class}
\alias{POSIXctz-class}
\alias{POSIXctz}
%% \alias{coerce}

\title{POSIXctz S4 class}
\description{NA}
\section{Objects from the Class}{Objects can be created by calls of the form \code{new(POSIXctz ...)}}
\section{Methods}{
  \describe{
    \item{coerce}{\code{signature(from = "character", to = "POSIXctz")}: ... }
	 }
}

\author{Thomas Wutzler}





