\name{seqRange}
\alias{seqRange}
\title{seqRange}
\description{Create a sequence based on a range c(min,max)}
\usage{seqRange(range, ...)}
\arguments{
  \item{range}{the range for the sequence}
  \item{\dots}{further arguments to seq, defaults to length.out=50}
}



\author{Thomas Wutzler}





