\name{setAlpha}
\alias{setAlpha}
\title{setAlpha}
\description{Setting a new alpha value (0..1 transparency) for color}
\usage{setAlpha(cols, alpha = 0.8)}
\arguments{
  \item{cols}{vector of colors}
  \item{alpha}{new alpha-Value}
}

\value{vector of colors}

\author{Thomas Wutzler}





