\name{str3}
\alias{str3}
\title{str3}
\description{Compactly Display the Structure with 3 nesting levels and without attributes}
\usage{
str3(object,
...)
}
\arguments{
  \item{object}{the object to display}
  \item{\dots}{further arguments to str}
}

\value{result of \code{\link{str}( object, max.level=3,  give.attr=FALSE, ...)}}

\author{Thomas Wutzler}



\seealso{\code{\link{copy2clip}}, \link{twMisc}}

