\name{traceback.curr}
\alias{traceback.curr}
\title{traceback curr}
\description{prints the current function calls, with the deepest call on top}
\usage{
traceback.curr(collapse = NULL)
}
\arguments{
  \item{collapse}{an optional character string to separate the results. Not NA_character_.}
}
\details{if collapse is given, nothing is printed but instead a string returned		}


\author{Thomas Wutzler}





