\name{twContCols16}
\alias{twContCols16}
\title{twContCols16}
\description{Return a character vector of green to magenta colorblind safe continuos scale}
\usage{twContCols16()}
\details{see  http://geography.uoregon.edu/datagraphics/color_scales.htm}


\author{Thomas Wutzler}



\seealso{\code{\link{twXlscol07}}, \code{\link{twMisc}}}

