\name{twDf2wikiTable}
\alias{twDf2wikiTable}
\title{twDf2wikiTable}
\description{Mediawiki-Code for table of given data.frame. }
\usage{
twDf2wikiTable(ds,
tableProps = "style=\\"float: right;\\"")
}
\arguments{
  \item{ds}{data.frame}
  \item{tableProps}{additional text inlcuded in the header}
}

\value{String, Side-effect: copied to clipboard}

\author{Thomas Wutzler}



\seealso{\code{\link{copy2clip}}, \link{twMisc}}

