\name{twDiffrl}
\alias{twDiffrl}
\title{twDiffrl}
\description{Approximation of dx by (x_+1 - x_-1). }
\usage{twDiffrl(x, na.rm = TRUE)}
\arguments{
  \item{x}{
}
  \item{na.rm}{can be set to FALSE for performance reasons}
}

\value{(x_+1 - x_-1)
with single sided intervals at the right and left side}

\author{Thomas Wutzler}





