\name{twEnumNames}
\alias{twEnumNames}
\title{twEnumNames}
\description{Creates integer variables with positions from colnames or names of x.}
\usage{
twEnumNames(x, inherits = TRUE, prefix = "i_")
}
\arguments{
  \item{x}{
}
  \item{inherits}{if FALSE checks only for existence in calling frame. Use with care!}
  \item{prefix}{string scalar: prefix for the variabl names}
}
\details{Use with care: new variables are defined in }
\value{Names of the variables assigned in parent frame.}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





