\name{twExtractDim}
\alias{twExtractDim}
\title{twExtractDim}
\description{Extract A[...,i,...] from the iDim dimension}
\usage{
twExtractDim(A, i = 1, iDim = length(dim(A)))
}
\arguments{
  \item{A}{the array to extract values from}
  \item{i}{the index to extract}
  \item{iDim}{the dimension to extract from}
}

\value{vector representing the }

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





