\name{twListArrDim}
\alias{twListArrDim}
\title{twListArrDim}
\description{Splits a dimension of an array to a list. (useful for do.call and apply)}
\usage{
twListArrDim(x, iDim = length(dim(x)))
}
\arguments{
  \item{x}{the array to split}
  \item{iDim}{the dimension to split along, defautls to last dimension}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





