\name{twMisc-package}
\alias{twMisc-package}\alias{twMisc}
\title{TWs development utilities}
\description{Utilities used during package development.}
\details{\describe{\item{Functionality of package twMisc}{ \itemize{\item{ create a sequence based on a range c(min,max): \code{\link{seqRange}} }\item{ merge several sequences to a single sequence: \code{\link{twMergeSequences}} }\item{ recursively merge a named list or vector to another: \code{\link{twMergeLists}} }\item{ assign first variable of RData-file into a variable: \code{\link{loadAssign}} }\item{ reorder factor levels: \code{\link{reorderFactor}} }\item{ format to significant number of digits including trailing zeros: \code{\link{formatSig}} }\item{ adds or replaces value in a vector: \code{\link{vectorElements<-}} }\item{ extracting data frame collumn while keeping rownames: \code{\link{dfcol}} }\item{ Retrieving and stripping filename extension : \code{\link{fileExt}},\code{\link{twStripFileExt}} }\item{ Error handling based on classes: \code{\link{userError}} }\item{ Array slicing: \code{\link{twExtractDim}} }\item{ RUnit tests for: Number inside Intervals: \code{\link{checkInterval}} and same magnitude \code{\link{checkMagnitude}} }\item debugging: \code{\link{traceback.curr}} \item working with constants: \code{\link{twEnumNames}} \item support for working with ODEs: \code{\link{twIntegrateTS}} \item Optimizing a function where first argument is an index.: \code{\link{twBinOptimize.numeric}} \item plotting routines: \code{\link{twPairs}}\item matching by closest value: \code{\link{matchClosest}}\item dealing with summer/winter time: \code{\link{strptimeNoSummer}}\item extracting from arrays: \code{\link{twExtractDim}}\item plotting image with row and column names: \code{\link{imageNames}}\item evaluating a function over a density adapted 2D grid for plotting: \code{\link{twApply2DMesh}}\item annotating a plot with arrows and text: \code{\link{locateArrowText}}}}}}%details\author{Thomas Wutzler}

\keyword{ package }


