\name{twPlot2DFun.contour}
\alias{twPlot2DFun.contour}
\title{twPlot2DFun contour}
\description{Applying FUN over grid of x and y values plot as an image.}
\usage{
twPlot2DFun.contour(...)
}
\arguments{
  \item{\dots}{further arguments passed to \code{\link{filled.contour}}}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





