\name{twSaveZero}
\alias{twSaveZero}
\title{twSaveZero}
\description{Function going to zero for small x over several orders of magnitude}
\usage{twSaveZero(x, kM = 1e-10)}
\arguments{
  \item{x}{numeric vector}
  \item{kM}{order of magnitude where decrease happens}
}
\details{f(x)->1 for x >> kM; f(x)=0 for x << kM }


\author{Thomas Wutzler}





