\name{twStripFileExt}
\alias{twStripFileExt}
\title{twStripFileExt}
\description{Remove the all the file extension, i.e. the last dot and suceeding characters.}
\usage{
twStripFileExt(filenames)
}
\arguments{
  \item{filenames}{
}
}



\author{Thomas Wutzler}



\seealso{code{\link{fileExt}}, \link{twMisc}}

