\name{whichColsEqualSumHeuristics}
\alias{whichColsEqualSumHeuristics}
\title{whichColsEqualSumHeuristics}
\description{compares each column of X to column of Z and returns indices of equal columns (determined by same sum column sum)    }
\usage{whichColsEqualSumHeuristics(X, Z)}
\arguments{
  \item{X}{matrix}
  \item{Z}{Matrix with same numer of rows, or a vector representing a single column that is compared to each column of X}
}



\author{Thomas Wutzler}





