\name{as.data.frame.RUnitTestData}
\alias{as.data.frame.RUnitTestData}
\title{as data frame RUnitTestData}
\description{extract information from RUnit Testresult return value}
\usage{
\method{as.data.frame}{RUnitTestData}(x,
row.names,
optional,
...)
}
\arguments{
  \item{x}{
}
  \item{row.names}{
}
  \item{optional}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{twUtestF}}}

