\name{copy2clip}
\alias{copy2clip}
\title{copy2clip}
\description{copies argument to the clipboard}
\usage{
copy2clip(x,
col.names = NA,
row.names = FALSE,
quote = FALSE,
sep = "\\t",
...)
}
\arguments{
  \item{x}{
}
  \item{col.names}{
}
  \item{row.names}{
}
  \item{quote}{
}
  \item{sep}{
}
  \item{\dots}{
}
}
\details{\describe{\item{Further misc functionality of package twMisc}{
\itemize{
\item{ easy copying to clipboard: this method }
\item{ create a sequence based on a range c(min,max): \code{\link{seqRange}} }
\item{ merge several sequences to a single sequence: \code{\link{twMergeSequences}} }
\item{ recursively merge a named list or vector to another: \code{\link{twMergeLists}} }
\item{ Mediawiki-Code for table of given data.frame: \code{\link{twDf2wikiTable}} }
\item{ assign first variable of RData-file into a variable: \code{\link{loadAssign}} }
\item{ reorder factor levels: \code{\link{reorderFactor}} }
\item{ format to significant number of digits including trailing zeros: \code{\link{formatSig}} }
\item{ adds or replaces value in a vector: \code{\link{vectorElements<-}} }
\item{ extracting data frame collumn while keepign rownames: \code{\link{dfcol}} }
\item{ Compactly Display the Structure (\code{str}) with 3 levels without attributes: \code{\link{str3}} }
\item{ TODO: link functions: \code{\link{twDf2wikiTable}} }
}
}}}


\author{Thomas Wutzler}



\seealso{\link{twMisc}}

