\name{copy2clip}
\alias{copy2clip}
\title{copy2clip}
\description{copies argument to the clipboard}
\usage{
copy2clip(x, col.names = NA, row.names = FALSE, quote = FALSE, ...)
}
\arguments{
  \item{x}{
}
  \item{col.names}{
}
  \item{row.names}{
}
  \item{quote}{
}
  \item{\dots}{
}
}
\details{\describe{\item{Further mics Functionality of package twMisc}{
\itemize{
\item{ easy copying to clipboard: this method }
\item{ Mediawiki-Code for table of given data.frame: \code{\link{twDf2wikiTable}} }
\item{ TODO: link functions: \code{\link{twDf2wikiTable}} }
}
}}}


\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\link{twMisc}}

