\name{seqRange}
\alias{seqRange}
\alias{twMisc}
\title{seqRange}
\description{Create a sequence based on a range c(min,max)}
\usage{
seqRange(range, ...)
}
\arguments{
  \item{range}{the range for the sequence}
  \item{\dots}{further arguments to seq, defaults to length.out=50}
}
\details{\describe{\item{Functionality of package twMisc}{
\itemize{
\item enhanced support for Unit-Tests: \code{\link{twUtestF}} 
\item debugging: \code{\link{traceback.curr}} 
\item working with constants: \code{\link{twEnumNames}} 
\item support for working with ODEs: \code{\link{twIntegrateTS}} 
\item Optimizing a function where first argument is an index.: \code{\link{twBinOptimize.numeric}} 
\item plotting routines: \code{\link{twPairs}}
\item matching by closest value: \code{\link{matchClosest}}
\item dealing with summer/winter time: \code{\link{strptimeNoSummer}}
\item collection of misc: \code{\link{copy2clip}}
\item extracting from arrays: \code{\link{twExtractDim}}
}
}}}


\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





