\name{seqRange}
\alias{seqRange}
\alias{twMisc}
\title{seqRange}
\description{Create a sequence based on a range c(min,max)}
\usage{
seqRange(range, ...)
}
\arguments{
  \item{range}{the range for the sequence}
  \item{\dots}{further arguments to seq, defaults to length.out=50}
}
\details{\describe{\item{Functionality of package twMisc}{
\itemize{
\item{ enhanced support for Unit-Tests: \code{\link{twUtestF}} }
\item{ debugging: \code{\link{traceback.curr}} }
\item{ working with constants: \code{\link{twEnumNames}} }
\item{ support for working with ODEs: \code{\link{twIntegrateTS}} }
\item{ Optimizing a function where first argument is an index.: \code{\link{twBinOptimize.numeric}} }
\item{ collection of misc }
}
}}}


\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





