\name{twExtractFromLastDims}
\alias{twExtractFromLastDims}
\title{twExtractFromLastDims}
\description{Extract slices i from array Aext keeping the all first dimensions.}
\usage{
twExtractFromLastDims(Aext, i, dPrev = c(1))
}
\arguments{
  \item{Aext}{array to extract from}
  \item{i}{indices in matrix of last dimensions}
  \item{dPrev}{dimensions in front to keep, defaults to rows}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





