\name{twLogMeanExp}
\alias{twLogMeanExp}
\title{twLogMeanExp}
\description{calculates the log(mean(exp(x))) in numerically safer way}
\usage{twLogMeanExp(x, ...)}
\arguments{
  \item{x}{vector for whose mean is to be caluclated}
  \item{\dots}{futher arguments to twLogSumExp}
}



\author{Thomas Wutzler}





