\name{twMisc-package}
\alias{twMisc-package}\alias{twMisc}
\title{TWs development utilities}
\description{Utilities used during package development.}
\details{\describe{\item{Functionality of package twMisc}{ \itemize{\item{ create a sequence based on a range c(min,max): \code{\link{seqRange}} }\item{ merge several sequences to a single sequence: \code{\link{twMergeSequences}} }\item{ recursively merge a named list or vector to another: \code{\link{twMergeLists}} }\item{ assign first variable of RData-file into a variable: \code{\link{loadAssign}} }\item{ reorder factor levels: \code{\link{reorderFactor}} }\item{ format to significant number of digits including trailing zeros: \code{\link{formatSig}} }\item{ adds or replaces value in a vector: \code{\link{vectorElements<-}} }\item{ extracting data frame collumn while keeping rownames: \code{\link{dfcol}} }\item{ Retrieving and stripping filename extension : \code{\link{fileExt}},\code{\link{twStripFileExt}} }\item{ Error handling based on classes: \code{\link{userError}} }\item{ Array slicing: \code{\link{twExtractDim}} }}}}}%details\author{Thomas Wutzler}

\keyword{ package }


