\name{twMisc-package}
\alias{twMisc-package}
\docType{package}
\title{tw utility functions}
\description{utilities used across several other packages of tw}
\details{
\tabular{ll}{Package: \tab twMisc\cr
Title: \tab tw utility functions\cr
Version: \tab 0.5-1\cr
Date: \tab 2010-04-15\cr
Author: \tab Thomas Wutzler\cr
Maintainer: \tab Thomas Wutzler <twutz@bgc-jena.mpg.de>\cr
Depends: \tab RUnit, R.methodsS3, abind\cr
Suggests: \tab inlinedocs,debug\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr}
}
\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}

\keyword{ package }


