\name{vectorElements<-}
\alias{vectorElements<-}
\title{vectorElements<-}
\description{adds or replaces value in vector}
\usage{
vectorElements<-(vec,
value)
}
\arguments{
  \item{vec}{a named vector}
  \item{value}{a named vector of same mode as vec}
}



\author{Thomas Wutzler}



\seealso{\code{\link{copy2clip}}, \link{twMisc}}

